feat: added navigation bar
