+++ app/src/main/java/com/yogeshpaliyal/deepr/ui/screens/Settings.kt
@@ -1,0 +1,55 @@
+package com.yogeshpaliyal.deepr.ui.screens
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.consumeWindowInsets
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.snapshots.SnapshotStateList
+import androidx.compose.ui.Modifier
+import compose.icons.TablerIcons
+import compose.icons.tablericons.ArrowBack
+import compose.icons.tablericons.ArrowLeft
+
+data object Settings
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SettingsScreen(backStack: SnapshotStateList<Any>) {
+    Scaffold(modifier = Modifier.fillMaxSize(), topBar = {
+        Column {
+            TopAppBar(
+                title = {
+                    Text("Settings")
+                },
+                navigationIcon = {
+                    IconButton(onClick = {
+                        backStack.removeLastOrNull()
+                    }) {
+                        Icon(
+                            TablerIcons.ArrowLeft,
+                            contentDescription = "Back"
+                        )
+                    }
+                }
+            )
+
+        }
+    }) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .padding(innerPadding)
+                .consumeWindowInsets(innerPadding)
+                .imePadding()
+        ) {
+            Text("Settings")
+        }
+    }
+}
